Fix integer overflow for TimeRemainingSeconds
- Change remainingSeconds from Int to Long in API response models
- Add capRemainingSeconds() to handle year-9999 expiry (max 30 days)
- Prevents Integer.MIN_VALUE when API returns very large values

+++ app/src/main/java/com/example/s7opcuaapp/data/api/LockModels.kt
@@ -15,7 +15,7 @@ data class LockStatusResponse(
     @SerializedName("DisplayName") val displayName: String? = null,
     @SerializedName("AcquiredAt") val acquiredAt: String? = null,
     @SerializedName("ExpiresAt") val expiresAt: String? = null,
-    @SerializedName("TimeRemainingSeconds") val remainingSeconds: Int? = null,
+    @SerializedName("TimeRemainingSeconds") val remainingSeconds: Long? = null,
     @SerializedName("IsCurrentUser") val isMyLock: Boolean = false
 )
 
@@ -32,7 +32,7 @@ data class AcquireLockRequest(
 data class AcquireLockResponse(
     @SerializedName("Success") val success: Boolean,
     @SerializedName("ExpiresAt") val expiresAt: String?,
-    @SerializedName("TimeRemainingSeconds") val remainingSeconds: Int?,
+    @SerializedName("TimeRemainingSeconds") val remainingSeconds: Long?,
     @SerializedName("Error") val error: String?
 )
 
@@ -57,7 +57,7 @@ data class ExtendLockRequest(
 data class ExtendLockResponse(
     @SerializedName("Success") val success: Boolean,
     @SerializedName("NewExpiresAt") val newExpiresAt: String?,
-    @SerializedName("TimeRemainingSeconds") val remainingSeconds: Int?,
+    @SerializedName("TimeRemainingSeconds") val remainingSeconds: Long?,
     @SerializedName("Error") val error: String?
 )
